requirements.txt
Django==3.0.3
psycopg2-binary==2.8.5
channels
django-cors-headers
djangorestframework